requirements.txt
requests
beautifulsoup4
tqdm
fpdf2
Pillow